refactor: extract shared name-parsing helpers into nameUtils
toLastFirstFormat / splitNames / parseAuthorName lived in metaTags.ts,
and bibtexGenerator independently re-implemented the exact splitNames
regex+trim+filter for editor strings. Move the three pure helpers to
nameUtils.ts; metaTags re-exports them (so its consumers are
unchanged) and bibtexGenerator now calls splitNames. Function bodies
are byte-identical, so no behavior change.
