+++ src/lib/utils/metaTags.ts
@@ -3,6 +3,12 @@
  * Provides common meta tag creation, author name formatting, and tag validation.
  */
 
+import { toLastFirstFormat } from '$lib/utils/nameUtils';
+
+// Name-string helpers live in nameUtils; re-exported here so existing
+// `$lib/utils/metaTags` imports keep working unchanged.
+export { toLastFirstFormat, splitNames, parseAuthorName } from '$lib/utils/nameUtils';
+
 export interface MetaTag {
 	name: string;
 	content: string;
@@ -18,29 +24,6 @@ export function createConditionalTag(
 ): MetaTag[] {
 	if (!content || !condition) return [];
 	return [{ name, content }];
-}
-
-/**
- * Converts "First Last" to "Last, First" format for citation managers (Zotero).
- * Names already in "Last, First" format are returned as-is.
- */
-export function toLastFirstFormat(name: string): string {
-	if (name.includes(',')) return name;
-	const parts = name.trim().split(/\s+/);
-	if (parts.length <= 1) return name;
-	const lastName = parts[parts.length - 1];
-	const firstName = parts.slice(0, -1).join(' ');
-	return `${lastName}, ${firstName}`;
-}
-
-/**
- * Splits a comma/and-separated name string into individual names.
- */
-export function splitNames(names: string): string[] {
-	return names
-		.split(/,\s*|\s+and\s+/)
-		.map((name) => name.trim())
-		.filter(Boolean);
 }
 
 /**
@@ -100,23 +83,4 @@ export function createCoinsParams(rfrId: string = 'info:sid/frederickmadore.com'
 	params.set('ctx_ver', 'Z39.88-2004');
 	params.set('rfr_id', rfrId);
 	return params;
-}
-
-/**
- * Parses an author name into first/last parts for COinS metadata.
- * Handles both "Last, First" and "First Last" formats.
- */
-export function parseAuthorName(author: string): { first?: string; last?: string } {
-	if (author.includes(',')) {
-		const [last, first] = author.split(',').map((s) => s.trim());
-		return { first, last };
-	}
-	const parts = author.trim().split(/\s+/);
-	if (parts.length > 1) {
-		return {
-			last: parts[parts.length - 1],
-			first: parts.slice(0, -1).join(' ')
-		};
-	}
-	return { last: author };
 }
